// src/app/lib/types.ts
export type WorldBankEntry = {
  indicator: {
    id: string;
    value: string;
  };
  country: {
    id: string;
    value: string;
  };
  countryiso3code: string;
  date: string;
  value: number | null;
  unit: string;
  obs_status: string;
  decimal: number;
};

export type PIBDataPoint = {
  year: string;
  PIB: number;
};